requirements.txt
fastapi
uvicorn
motor
python-dotenv
fastapi-mail
pydantic
gunicorn
uvicorn[standard]
